crawler/requirements.txt
scrapy>=2.11
lxml>=5.0.0
psycopg2-binary>=2.9.9
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
